refactor: factored out MainWindow
